Rename beans and services for consistency
