+++ src/main/java/de/aservo/atlassian/confapi/model/MailServerSmtpBean.java
@@ -1,6 +1,6 @@
 package de.aservo.atlassian.confapi.model;
 
-import com.atlassian.mail.server.PopMailServer;
+import com.atlassian.mail.server.SMTPMailServer;
 import de.aservo.atlassian.confapi.constants.ConfAPI;
 import de.aservo.atlassian.confapi.exception.NoContentException;
 import org.apache.commons.lang3.StringUtils;
@@ -15,16 +15,25 @@ import javax.xml.bind.annotation.XmlRootElement;
 import static com.atlassian.mail.MailConstants.DEFAULT_TIMEOUT;
 
 /**
- * Bean for POP mail server in REST requests.
+ * Bean for SMTP mail server in REST requests.
  */
-@XmlRootElement(name = ConfAPI.MAIL_SERVER_POP)
-public class PopMailServerBean {
+@XmlRootElement(name = ConfAPI.MAIL_SERVER_SMTP)
+public class MailServerSmtpBean {
 
     @XmlElement
     private final String name;
 
     @XmlElement
     private final String description;
+
+    @XmlElement
+    private final String adminContact;
+
+    @XmlElement
+    private final String from;
+
+    @XmlElement
+    private final String prefix;
 
     @XmlElement
     private final String protocol;
@@ -34,6 +43,9 @@ public class PopMailServerBean {
 
     @XmlElement
     private final Integer port;
+
+    @XmlElement
+    private final boolean tls;
 
     @XmlElement
     private final long timeout;
@@ -49,31 +61,43 @@ public class PopMailServerBean {
     /**
      * The default constructor is needed for JSON request deserialization.
      */
-    public PopMailServerBean() {
+    public MailServerSmtpBean() {
         this.name = null;
         this.description = null;
+        this.adminContact = null;
+        this.from = null;
+        this.prefix = null;
         this.protocol = null;
         this.host = null;
         this.port = null;
+        this.tls = false;
         this.timeout = DEFAULT_TIMEOUT;
         this.username = null;
         this.password = null;
     }
 
-    private PopMailServerBean(
+    public MailServerSmtpBean(
             final String name,
             final String description,
+            final String adminContact,
+            final String from,
+            final String prefix,
             final String protocol,
             final String host,
             final Integer port,
+            final boolean tls,
             final long timeout,
             final String username) {
 
         this.name = name;
         this.description = StringUtils.isNoneBlank(description) ? description : null;
+        this.adminContact = adminContact;
+        this.from = from;
+        this.prefix = prefix;
         this.protocol = protocol;
         this.host = host;
         this.port = port;
+        this.tls = tls;
         this.timeout = timeout;
         this.username = username;
         this.password = null;
@@ -85,6 +109,18 @@ public class PopMailServerBean {
 
     public String getDescription() {
         return description;
+    }
+
+    public String getAdminContact() {
+        return adminContact;
+    }
+
+    public String getFrom() {
+        return from;
+    }
+
+    public String getPrefix() {
+        return prefix;
     }
 
     public String getProtocol() {
@@ -99,6 +135,10 @@ public class PopMailServerBean {
         return port;
     }
 
+    public boolean isTls() {
+        return tls;
+    }
+
     public long getTimeout() {
         return timeout;
     }
@@ -111,21 +151,25 @@ public class PopMailServerBean {
         return password;
     }
 
-    public static PopMailServerBean from(
-            final PopMailServer popMailServer) throws NoContentException {
+    public static MailServerSmtpBean from(
+            final SMTPMailServer smtpMailServer) throws NoContentException {
 
-        if (popMailServer == null) {
-            throw new NoContentException("No POP mail server defined");
+        if (smtpMailServer == null) {
+            throw new NoContentException("No SMTP mail server defined");
         }
 
-        return new PopMailServerBean(
-                popMailServer.getName(),
-                popMailServer.getDescription(),
-                popMailServer.getMailProtocol().getProtocol(),
-                popMailServer.getHostname(),
-                StringUtils.isNotBlank(popMailServer.getPort()) ? Integer.parseInt(popMailServer.getPort()) : null,
-                popMailServer.getTimeout(),
-                popMailServer.getUsername());
+        return new MailServerSmtpBean(
+                smtpMailServer.getName(),
+                smtpMailServer.getDescription(),
+                null,
+                smtpMailServer.getDefaultFrom(),
+                smtpMailServer.getPrefix(),
+                smtpMailServer.getMailProtocol().getProtocol(),
+                smtpMailServer.getHostname(),
+                StringUtils.isNotBlank(smtpMailServer.getPort()) ? Integer.parseInt(smtpMailServer.getPort()) : null,
+                smtpMailServer.isTlsRequired(),
+                smtpMailServer.getTimeout(),
+                smtpMailServer.getUsername());
     }
 
     @Override
