+++ src/main/java/de/aservo/atlassian/confapi/service/DirectoryService.java
@@ -1,21 +1,21 @@
 package de.aservo.atlassian.confapi.service;
 
 import com.atlassian.crowd.exception.DirectoryCurrentlySynchronisingException;
-import de.aservo.atlassian.confapi.model.UserDirectoryBean;
+import de.aservo.atlassian.confapi.model.DirectoryBean;
 
 import java.util.List;
 
 /**
  * The User directory service interface.
  */
-public interface UserDirectoryService {
+public interface DirectoryService {
 
     /**
      * Gets directories.
      *
      * @return the directories
      */
-    List<UserDirectoryBean> getDirectories();
+    List<DirectoryBean> getDirectories();
 
     /**
      * Adds a new directory configurations. Any existing configurations with the same 'name' property are removed before adding the new configuration.
@@ -25,5 +25,5 @@ public interface UserDirectoryService {
      * @return the configuration added
      * @throws DirectoryCurrentlySynchronisingException the directory currently synchronising exception
      */
-    UserDirectoryBean addDirectory(UserDirectoryBean directory, boolean testConnection) throws DirectoryCurrentlySynchronisingException;
+    DirectoryBean addDirectory(DirectoryBean directory, boolean testConnection) throws DirectoryCurrentlySynchronisingException;
 }
