+++ src/test/java/de/aservo/atlassian/confapi/model/MailServerPopBeanTest.java
@@ -1,29 +1,27 @@
 package de.aservo.atlassian.confapi.model;
 
-import com.atlassian.mail.server.DefaultTestSmtpMailServerImpl;
-import com.atlassian.mail.server.SMTPMailServer;
+import com.atlassian.mail.server.DefaultTestPopMailServerImpl;
+import com.atlassian.mail.server.PopMailServer;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.junit.MockitoJUnitRunner;
 
 import static com.atlassian.mail.MailConstants.DEFAULT_TIMEOUT;
-import static org.junit.Assert.*;
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertNull;
 
 @RunWith(MockitoJUnitRunner.class)
-public class SmtpMailServerBeanTest {
+public class MailServerPopBeanTest {
 
     @Test
     public void testDefaultConstructor() {
-        final SmtpMailServerBean bean = new SmtpMailServerBean();
+        final MailServerPopBean bean = new MailServerPopBean();
 
         assertNull(bean.getName());
         assertNull(bean.getDescription());
-        assertNull(bean.getFrom());
-        assertNull(bean.getPrefix());
         assertNull(bean.getProtocol());
         assertNull(bean.getHost());
         assertNull(bean.getPort());
-        assertFalse(bean.isTls());
         assertEquals(DEFAULT_TIMEOUT, bean.getTimeout());
         assertNull(bean.getUsername());
         assertNull(bean.getPassword());
@@ -31,17 +29,14 @@ public class SmtpMailServerBeanTest {
 
     @Test
     public void testFromConstructor() throws Exception {
-        final SMTPMailServer server = new DefaultTestSmtpMailServerImpl();
-        final SmtpMailServerBean bean = SmtpMailServerBean.from(server);
+        final PopMailServer server = new DefaultTestPopMailServerImpl();
+        final MailServerPopBean bean = MailServerPopBean.from(server);
 
         assertEquals(server.getName(), bean.getName());
         assertEquals(server.getDescription(), bean.getDescription());
-        assertEquals(server.getDefaultFrom(), bean.getFrom());
-        assertEquals(server.getPrefix(), bean.getPrefix());
         assertEquals(server.getMailProtocol().getProtocol(), bean.getProtocol());
         assertEquals(server.getHostname(), bean.getHost());
         assertEquals(Integer.valueOf(server.getPort()), bean.getPort());
-        assertEquals(server.isTlsRequired(), bean.isTls());
         assertEquals(server.getTimeout(), bean.getTimeout());
         assertEquals(server.getUsername(), bean.getUsername());
         assertNull(bean.getPassword());
@@ -49,9 +44,9 @@ public class SmtpMailServerBeanTest {
 
     @Test
     public void testFromConstructorHideEmptyDescription() throws Exception {
-        final SMTPMailServer server = new DefaultTestSmtpMailServerImpl();
+        final PopMailServer server = new DefaultTestPopMailServerImpl();
         server.setDescription("");
-        final SmtpMailServerBean bean = SmtpMailServerBean.from(server);
+        final MailServerPopBean bean = MailServerPopBean.from(server);
 
         assertNull(bean.getDescription());
     }
