+++ src/test/java/de/aservo/atlassian/confapi/model/MailServerSmtpBeanTest.java
@@ -1,27 +1,29 @@
 package de.aservo.atlassian.confapi.model;
 
-import com.atlassian.mail.server.DefaultTestPopMailServerImpl;
-import com.atlassian.mail.server.PopMailServer;
+import com.atlassian.mail.server.DefaultTestSmtpMailServerImpl;
+import com.atlassian.mail.server.SMTPMailServer;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.junit.MockitoJUnitRunner;
 
 import static com.atlassian.mail.MailConstants.DEFAULT_TIMEOUT;
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertNull;
+import static org.junit.Assert.*;
 
 @RunWith(MockitoJUnitRunner.class)
-public class PopMailServerBeanTest {
+public class MailServerSmtpBeanTest {
 
     @Test
     public void testDefaultConstructor() {
-        final PopMailServerBean bean = new PopMailServerBean();
+        final MailServerSmtpBean bean = new MailServerSmtpBean();
 
         assertNull(bean.getName());
         assertNull(bean.getDescription());
+        assertNull(bean.getFrom());
+        assertNull(bean.getPrefix());
         assertNull(bean.getProtocol());
         assertNull(bean.getHost());
         assertNull(bean.getPort());
+        assertFalse(bean.isTls());
         assertEquals(DEFAULT_TIMEOUT, bean.getTimeout());
         assertNull(bean.getUsername());
         assertNull(bean.getPassword());
@@ -29,14 +31,17 @@ public class PopMailServerBeanTest {
 
     @Test
     public void testFromConstructor() throws Exception {
-        final PopMailServer server = new DefaultTestPopMailServerImpl();
-        final PopMailServerBean bean = PopMailServerBean.from(server);
+        final SMTPMailServer server = new DefaultTestSmtpMailServerImpl();
+        final MailServerSmtpBean bean = MailServerSmtpBean.from(server);
 
         assertEquals(server.getName(), bean.getName());
         assertEquals(server.getDescription(), bean.getDescription());
+        assertEquals(server.getDefaultFrom(), bean.getFrom());
+        assertEquals(server.getPrefix(), bean.getPrefix());
         assertEquals(server.getMailProtocol().getProtocol(), bean.getProtocol());
         assertEquals(server.getHostname(), bean.getHost());
         assertEquals(Integer.valueOf(server.getPort()), bean.getPort());
+        assertEquals(server.isTlsRequired(), bean.isTls());
         assertEquals(server.getTimeout(), bean.getTimeout());
         assertEquals(server.getUsername(), bean.getUsername());
         assertNull(bean.getPassword());
@@ -44,9 +49,9 @@ public class PopMailServerBeanTest {
 
     @Test
     public void testFromConstructorHideEmptyDescription() throws Exception {
-        final PopMailServer server = new DefaultTestPopMailServerImpl();
+        final SMTPMailServer server = new DefaultTestSmtpMailServerImpl();
         server.setDescription("");
-        final PopMailServerBean bean = PopMailServerBean.from(server);
+        final MailServerSmtpBean bean = MailServerSmtpBean.from(server);
 
         assertNull(bean.getDescription());
     }
